fix: ics download doesn't require oauth anymore, still require formatting
