+++ src/popup/popup.html
@@ -15,11 +15,11 @@
             </button>
         </div>
         <div class="middle">
-            <img src="../images/magnify128.png" width=30px height=auto></img>
+            <img src="../../images/magnify128.png" width=30px height=auto></img>
             <h1>Schedulr</h1>
         </div>
         <a href="https://hackerspacemmu.rocks/" target="_blank">
-            <img src="../images/hackerspace.png" width=40px height=auto></img>
+            <img src="../../images/hackerspace.png" width=40px height=auto></img>
         </a>
     </div>
 
@@ -122,5 +122,8 @@
         <input type="submit" value="Submit">
     </div>
     <script type="module" src="popup.js"></script>
+    <script type="module" src="../scripts/edit-cal-choice.js"></script>
+    <script type="module" src="../scripts/btn-behaviour.js"></script>
+    <script type="module" src="../scripts/field-checker.js"></script>
 </body>
 </html>
