+++ src/popup/popup.js
@@ -1,165 +1,20 @@
 import { getAuthToken, getCurrTab } from '../scripts/helper/prog-flow.js';
 
 // Query for user available calendar then insert into popup.html dynamically
-function queryCalList() {
-    chrome.runtime.sendMessage({
-        action: "queryCalList"
-    });
-}
-
-queryCalList();
-
-chrome.runtime.onMessage.addListener((message) => {
-    if (message.action === "calData") {
-        let calData = message.data;
-        console.log("Cal Data:", calData);
-
-        // Get form element in html
-        const form = document.getElementById("calendarForm");
-
-        setAttributes(form, calData);
-    } else if (message.action === "showAlert") {
-        window.alert(calData);
-    }
+chrome.runtime.sendMessage({
+    action: "queryCalList"
 });
-
-function setAttributes(form, calData) {
-    for (let cals in calData) {
-        // Create input and label tag for every index
-        const input = document.createElement("input");
-        const label = document.createElement("label");
-        const br = document.createElement("br");
-
-        // Set attribute for input tag
-        input.setAttribute("type", "radio");
-        input.setAttribute("id", `${cals}`);
-        input.setAttribute("name", "calendar");
-        input.setAttribute("value", `${calData[cals]}`);
-
-        // Set attribute for label tag
-        label.innerText = `${cals}`;
-        label.setAttribute("for", `${cals}`);
-
-        // Append input and label tag, then a line break after
-        form.appendChild(input);
-        form.appendChild(label);
-        form.appendChild(br);
-    }
-
-    // Hide the loader after calendar
-    document.getElementById('loader').style.display = 'none';
-}
-
-let selectedOptionValue;
-
-// Function to handle event
-function finalButtonClickHandler(event) {
-    event.preventDefault();
-    getFormsValue(selectedOptionValue);
-}
-
-// Listener to get selected option value
-document.getElementById('optionForm').addEventListener('submit', function(event) {
-    event.preventDefault();
-
-    selectedOptionValue = document.querySelector('input[name="option"]:checked')?.value;
-    console.log(`Received option value: ${selectedOptionValue}`);
-
-    if (!(selectedOptionValue)) {
-        window.alert('Please select an option.');
-        return;
-    }
-
-    // Hide the loader after calendar
-    const firstPage = document.getElementsByClassName('firstPage')[0].style.display = 'none';
-
-    // Query all possible forms
-    let generalForms = document.getElementsByClassName('generalForms')[0];
-    let backButton = document.getElementById('backButton');
-    let calForms = document.getElementsByClassName('calForms')[0];
-    let finalButton = document.getElementsByClassName('finalButton')[0];
-
-    generalForms.style.display = 'flex';
-    backButton.style.display = 'flex';
-
-    // Display appropriate forms
-    if (selectedOptionValue == 1 || selectedOptionValue == 3) {
-        calForms.style.display = 'flex';
-        finalButton.style.display = 'flex';
-    } else if (selectedOptionValue == 2) {
-        finalButton.style.display = 'flex';
-    }
-
-    // Remove previously attached listener
-    finalButton.removeEventListener('click', finalButtonClickHandler);
-
-    // Attach new listener
-    finalButton.addEventListener('click', finalButtonClickHandler);
-});
-
-// Event listener for the back button
-document.getElementById('backButton').addEventListener('click', function() {
-    // Query all forms
-    let optionForm = document.getElementsByClassName('optionForm')[0];
-    let generalForms = document.getElementsByClassName('generalForms')[0];
-    let calForms = document.getElementsByClassName('calForms')[0];
-    let finalButton = document.getElementsByClassName('finalButton')[0];
-
-    // Hide the second page forms
-    generalForms.style.display = 'none';
-    calForms.style.display = 'none';
-    finalButton.style.display = 'none';
-
-    // Clear the previous form options
-    // optionForm.innerHTML = '';
-
-    // Show the first page again
-    document.getElementsByClassName('firstPage')[0].style.display = 'flex';
-
-    // Hide the back button itself since we're back on the first page
-    document.getElementById('backButton').style.display = 'none';
-});
-
-// Checks all neccessary field vary by selected option value
-function getFormsValue(selectedOptionValue) {
-    try{
-        // Get the value of the selected radio button
-        const selectedSemesterValue = document.querySelector('input[name="semester"]:checked')?.value;
-        const selectedReminderTime = document.querySelector('input[name="reminder"]:checked')?.value;
-        const selectedColorValue = document.querySelector('input[name="color"]:checked')?.value;
-        const selectedCalendar = document.querySelector('input[name="calendar"]:checked')?.value;
-        const selectedEventFormat = document.querySelector('input[name="format"]:checked')?.value;
-
-        if (selectedOptionValue == 1 || selectedOptionValue == 3) {
-            // Check if all values are selected
-            if (!(selectedSemesterValue && selectedReminderTime && selectedColorValue && selectedCalendar && selectedEventFormat)) {
-                window.alert('Please select all options.');
-                return;
-            }
-
-            handleFlow(selectedColorValue, selectedCalendar, selectedReminderTime, selectedSemesterValue, selectedEventFormat, selectedOptionValue);
-
-        } else if (selectedOptionValue == 2) {
-            // Check if all values are selected
-            if (!(selectedSemesterValue && selectedReminderTime && selectedEventFormat)) {
-                window.alert('Please select all options.');
-                return;
-            }
-
-            handleFlow(null, null, selectedReminderTime, selectedSemesterValue, selectedEventFormat, selectedOptionValue);
-        }
-    }
-    catch(err) {
-        console.error('An error occured: ', err);
-        window.alert(`An error occured: ${err.message}`);
-    }
-}
 
 // This function handles token and window flow
-async function handleFlow(selectedColorValue, selectedCalendar, selectedReminderTime, selectedSemesterValue, selectedEventFormat, selectedOptionValue) {
+export async function handleFlow(selectedColorValue, selectedCalendar, selectedReminderTime, selectedSemesterValue, selectedEventFormat, selectedOptionValue) {
     try {
-        // Get Oauth token
-        const token = await getAuthToken();
+        let token = null;
+
+        // Only try to get token if the selected options require Google account access
+        if (selectedOptionValue == 1 || selectedOptionValue == 3) {
+            // Get Oauth token
+            token = await getAuthToken();
+        }
 
         // Get the current active tab
         const currTab = await getCurrTab();
